Move RefInfo classes into opinfo.refs
Given that there is already a clear `op_db`, `python_ref_db` split I
think it makes sense to have the `RefInfo` classes be defined in a
different file.

[ghstack-poisoned]
